Python-Appium/requirement.txt

Appium-Python-Client==2.11.1
selenium==4.12.0
python-dotenv=1.0.0